Began major refactor
