+++ src/filesystem/ThetisFS.h
@@ -14,4 +14,4 @@ bool appendFile(fs::FS &fs, const char * path, const char * message);
 bool renameFile(fs::FS &fs, const char * path1, const char * path2);
 bool deleteFile(fs::FS &fs, const char * path);
 void testFileIO(fs::FS &fs, const char * path);
-bool initLogFile(fs::FS &fs, char * path, char * header);
+
